+++ lib/src/rust/frb_generated.dart
@@ -2931,7 +2931,7 @@ class LegadoEngineApiImpl extends LegadoEngineApiImplPlatform
         },
         codec: SseCodec(
           decodeSuccessData: sse_decode_unit,
-          decodeErrorData: sse_decode_String,
+          decodeErrorData: sse_decode_app_error,
         ),
         constMeta: kCrateApiSeedLoginHeaderConstMeta,
         argValues: [sourceUrl, header],